feat(plutus): Moved the PlutusV3 specific tests
Moved the PlutusV3 specific tests from `test_mint_build.py` to a new file
`test_mint_build.py` under the `tests_plutus_v3` directory. This change
organizes the tests better by separating PlutusV3 specific tests from
other Plutus tests.
